refactor: use just functions for toTransaction and toUtxo rather than class with static function
